+++ src/onegov.form/HISTORY.rst
@@ -1,5 +1,9 @@
 Changelog
 ---------
+
+- Adds a phone number validator.
+  [msom]
+
 0.35.0 (2018-09-12)
 ~~~~~~~~~~~~~~~~~~~